Reorganize WebIDL into a formal integration
