+++ webidl/typeconstructor/typeconstructor_test.go
@@ -14,7 +14,8 @@ import (
 	"github.com/serulian/compiler/compilergraph"
 	"github.com/serulian/compiler/graphs/typegraph"
 	"github.com/serulian/compiler/packageloader"
-	"github.com/serulian/compiler/webidl"
+	webidl "github.com/serulian/compiler/webidl/graph"
+
 	"github.com/stretchr/testify/assert"
 )
